LayoutTests/platform/gtk/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/mo-minsize-maxsize-legacy-length-values-expected.txt


PASS minsize does not accept nonzero unitless values
FAIL maxsize does not accept nonzero unitless values assert_approx_equals: expected 125 +/- 1 but got 25
FAIL maxsize does not accept namedspace values assert_approx_equals: expected 125 +/- 1 but got 25
⥯
⥯
⥯
⥯
⥯
⥯
⥯
⥯
⥯
⥯
⥯
⥯
⥯
⥯
⥯
⥯
